+++ common/spec/dependabot/metadata_finders/base/release_finder_spec.rb
@@ -115,8 +115,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
               expect(subject).
                 to eq(
                   "## v1.8.0\n"\
-                  "- Add 2018-2027 TARGET holiday defintions\n"\
-                  "- Add 2018-2027 Bankgirot holiday defintions"
+                  "- Add 2018-2027 TARGET holiday definitions\n"\
+                  "- Add 2018-2027 Bankgirot holiday definitions"
                 )
             end
 
@@ -162,8 +162,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
                 expect(subject).
                   to eq(
                     "## v1.8.0\n"\
-                    "- Add 2018-2027 TARGET holiday defintions\n"\
-                    "- Add 2018-2027 Bankgirot holiday defintions"
+                    "- Add 2018-2027 TARGET holiday definitions\n"\
+                    "- Add 2018-2027 Bankgirot holiday definitions"
                   )
               end
             end
@@ -177,8 +177,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
                 expect(subject).
                   to eq(
                     "## business-1.8.0\n"\
-                    "- Add 2018-2027 TARGET holiday defintions\n"\
-                    "- Add 2018-2027 Bankgirot holiday defintions"
+                    "- Add 2018-2027 TARGET holiday definitions\n"\
+                    "- Add 2018-2027 Bankgirot holiday definitions"
                   )
               end
             end
@@ -206,8 +206,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
                 expect(subject).
                   to eq(
                     "## v1.8.0\n"\
-                    "- Add 2018-2027 TARGET holiday defintions\n"\
-                    "- Add 2018-2027 Bankgirot holiday defintions"
+                    "- Add 2018-2027 TARGET holiday definitions\n"\
+                    "- Add 2018-2027 Bankgirot holiday definitions"
                   )
               end
             end
@@ -253,8 +253,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
               expect(subject).
                 to eq(
                   "## v1.8.0\n"\
-                  "- Add 2018-2027 TARGET holiday defintions\n"\
-                  "- Add 2018-2027 Bankgirot holiday defintions\n"\
+                  "- Add 2018-2027 TARGET holiday definitions\n"\
+                  "- Add 2018-2027 Bankgirot holiday definitions\n"\
                   "\n"\
                   "## v1.7.0\n"\
                   "No release notes provided.\n"\
@@ -305,8 +305,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
               expect(subject).
                 to eq(
                   "## v1.8.0\n"\
-                  "- Add 2018-2027 TARGET holiday defintions\n"\
-                  "- Add 2018-2027 Bankgirot holiday defintions\n"\
+                  "- Add 2018-2027 TARGET holiday definitions\n"\
+                  "- Add 2018-2027 Bankgirot holiday definitions\n"\
                   "\n"\
                   "## v1.7.0\n"\
                   "No release notes provided.\n"\
@@ -364,8 +364,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
             expect(subject).
               to eq(
                 "## v1.7.0\n"\
-                "- Add 2018-2027 TARGET holiday defintions\n"\
-                "- Add 2018-2027 Bankgirot holiday defintions"
+                "- Add 2018-2027 TARGET holiday definitions\n"\
+                "- Add 2018-2027 Bankgirot holiday definitions"
               )
           end
         end
@@ -418,8 +418,8 @@ RSpec.describe Dependabot::MetadataFinders::Base::ReleaseFinder do
               expect(subject).
                 to eq(
                   "## v1.8.0\n"\
-                  "- Add 2018-2027 TARGET holiday defintions\n"\
-                  "- Add 2018-2027 Bankgirot holiday defintions"
+                  "- Add 2018-2027 TARGET holiday definitions\n"\
+                  "- Add 2018-2027 Bankgirot holiday definitions"
                 )
             end
           end