+++ CHANGELOG.md
@@ -1615,7 +1615,7 @@
 ## v0.98.25, 21 March 2019
 
 - Python: Split Pipfile manifest updater into separate class
-- Use GitHub repo name defintion for GitLab and Azure
+- Use GitHub repo name definition for GitLab and Azure
 - PHP: Handle relative paths that are actually from the root
 - Rebuild Dockerfile using Ruby 2.6.2